danial

+++ application/modules/cities/controllers/RequestController.php
@@ -14,7 +14,7 @@ class Cities_RequestController extends MyIndo_Controller_Action
 	public function createAction()
 	{
 		try {
-			if(isset($this->_posts['NAME'])) {
+			if(isset($this->_posts['NAME']) && !empty($this->_posts['NAME'])) {
 				if(!$this->_model->isExist('NAME', $this->_posts['NAME'])) {
 					
 					$this->_model->insert(array(
@@ -35,10 +35,10 @@ class Cities_RequestController extends MyIndo_Controller_Action
 	public function readAction()
 	{
 		try {
-			if(isset($this->_posts['query'])) {
+			if(isset($this->_posts['query']) && !empty($this->_posts['query'])) {
 				$this->_where[] = $this->_model->getAdapter()->quoteInto('NAME LIKE ?', '%' . $this->_posts['query'] . '%');
 			}
-			if(isset($this->_posts['NAME'])) {
+			if(isset($this->_posts['NAME']) && !empty($this->_posts['NAME'])) {
 				$this->_where[] = $this->_model->getAdapter()->quoteInto('NAME LIKE ?', '%' . $this->_posts['NAME'] . '%');
 			}
 			$this->_data['items'] = $this->_model->getList($this->_limit, $this->_start, $this->_order, $this->_where);
@@ -51,7 +51,7 @@ class Cities_RequestController extends MyIndo_Controller_Action
 	public function updateAction()
 	{
 		try {
-			if(isset($this->_posts['ID']) && isset($this->_posts['NAME'])) {
+			if(isset($this->_posts['ID']) && isset($this->_posts['NAME']) && !empty($this->_posts['ID']) && !empty($this->_posts['NAME'])) {
 				$id = $this->_enc->base64decrypt($this->_posts['ID']);
 				$name = $this->_posts['NAME'];
 				$valid = true;
@@ -85,7 +85,7 @@ class Cities_RequestController extends MyIndo_Controller_Action
 	public function destroyAction()
 	{
 		try {
-			if(isset($this->_posts['ID'])) {
+			if(isset($this->_posts['ID']) && !empty($this->_posts['ID'])) {
 				$id = $this->_enc->base64decrypt($this->_posts['ID']);
 				if($this->_model->isExist('ID', $id)) {
 					$this->_model->delete($this->_model->getAdapter()->quoteInto('ID = ?', $id));
