+++ application/modules/trainingparticipants/controllers/RequestController.php
@@ -114,7 +114,7 @@ class Trainingparticipants_RequestController extends MyIndo_Controller_Action
 	public function readAction()
 	{
 		try {
-			if(isset($this->_posts['TRAINING_ID'])) {
+			if(isset($this->_posts['TRAINING_ID']) && !empty($this->_posts['TRAINING_ID'])) {
 				$trainingId = (int)$this->_enc->base64decrypt($this->_posts['TRAINING_ID']);
 				if($this->_model->isExist('TRAINING_ID', $trainingId)) {
 					$this->_where[] = $this->_modelView->getAdapter()->quoteInto('TRAINING_ID = ?', $trainingId);
@@ -196,7 +196,7 @@ class Trainingparticipants_RequestController extends MyIndo_Controller_Action
 	public function destroyAction()
 	{
 		try {
-			if(isset($this->_posts['ID'])) {
+			if(isset($this->_posts['ID']) && !empty($this->_posts['ID'])) {
 				$id = $this->_enc->base64decrypt($this->_posts['ID']);
 				if($this->_model->isExist('ID', $id)) {
 					$this->_model->delete($this->_model->getAdapter()->quoteInto('ID = ?', $id));
